DW-183 [Backend][Misc] cast to long according to sonarqube report

+++ src/main/java/de/digiwill/service/RegistrationService.java
@@ -93,7 +93,7 @@ public class RegistrationService {
                 personalData,
                 authorities);
         userHandle.sendSignOfLife();
-        userHandle.setDeltaDeathTime(60 * 60 * 24 * 14); //Sets time to 14 days
+        userHandle.setDeltaDeathTime((long)60 * 60 * 24 * 14); //Sets time to 14 days
         return userHandle;
     }
 }
